backend/requirements.txt
SQLAlchemy
Flask
boto3
PyMuPDF
psycopg2-binary
python-dotenv
flask_sqlalchemy
flask_cors
